added a base class for OpenAIFunctionsAgentOutputParser
